Enhancement: Proposal to pass multiple subfolder in a folder Fixes #29
Added capability to the parser to accept an array of strings as the folder directories. Make the parser go through all the given folders in the same way we used to go through one folder before. I also modified the CLI to allow for multiple directories.
Modified the Readme file to have this new information and regenerated the ClassDiagram since now we have added new elements to the testing folders.